+++ src/components/App.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from "react";
 import { BrowserRouter as Router, Routes, Route, Navigate, useLocation } from "react-router-dom";
 import { Helmet } from "react-helmet-async";
 
-import VertexView from "./routes/VertexView";
-import DagView from "./routes/DagExplorerView";
-import loadDatasetGraph from "./util/loadDatasetGraph";
-import { DatasetGraph, Ontology } from "./d";
+import VertexView from "../routes/VertexView";
+import DagView from "../routes/DagExplorerView";
+import loadDatasetGraph from "../util/loadDatasetGraph";
+import { DatasetGraph, Ontology } from "../d";
 
 interface AppState {
   graph?: DatasetGraph;
