+++ src/components/Controls.tsx
@@ -12,11 +12,11 @@ import {
 
 import { useParams } from "react-router-dom";
 
-import { OntologyTerm } from "../../d";
+import { OntologyTerm } from "../d";
 import { Link } from "react-router-dom";
-import { settingsDrawerActiveState } from "../../recoil/controls";
+import { settingsDrawerActiveState } from "../recoil/controls";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { sugiyamaIsEnabledState } from "../../recoil/sugi";
+import { sugiyamaIsEnabledState } from "../recoil/sugi";
 
 interface OntologyExplorerControlDrawerProps {
   pinnedVertex: OntologyTerm | undefined;
