+++ src/components/SearchSidebar.tsx
